Perform an immune cell deconvolution on a dataset.
Source: R/immune_deconvolution_methods.R

Perform an immune cell deconvolution on a dataset.

deconvolute.default(gene_expression, method = deconvolution_methods, ...)

deconvolute(gene_expression, method = deconvolution_methods,
  column = "gene_symbol")

# S4 method for eSet,character
deconvolute(gene_expression,
  method = deconvolution_methods, column = "gene_symbol")

# S4 method for matrix,character
deconvolute(gene_expression, method)

Arguments

gene_expression
A numeric matrix with HGNC gene symbols as rownames and sample identifiers as colnames. Data must be on non-log scale.

method
a string specifying the method. Supported methods are xcell, ...

...
arguments passed to the respective method

eset
Expression set

col
column name of the fData column that contains HGNC gene symbols.

Value

data.frame with cell_type as first column and a column with the calculated cell fractions for each sample.

Methods (by class)

gene_expression = eSet,method = character: eset is a Biobase::ExpressionSet. fData contains a column with HGNC gene symbols (specify column name)

gene_expression = matrix,method = character: matrix is matrix with HGNC gene symbols as row names.

Examples

# Not run: deconvolute(gene_expression_matrix, "xcell")