Deconvolute using CIBERSORT or CIBERSORT abs.

Deconvolute using CIBERSORT or CIBERSORT abs.

deconvolute_cibersort(gene_expression_matrix, arrays, absolute = FALSE,
  abs_method = "sig.score", ...)

Arguments

gene_expression_matrix
a m x n matrix with m genes and n samples

arrays
Set to TRUE if working with Microarray data instead of RNA-seq. As recommended by the authors, quantile normalization will be enabled for Microarrays and disabled for RNAseq.

absolute
Set to TRUE for CIBERSORT absolute mode.

abs_method
Choose method to compute absolute score (only if absolute=TRUE).

...
passed through to the original CIBERSORT function. A native argument takes precedence over an immunedeconv argument (e.g. QN takes precedence over arrays). Documentation is not publicly available. Log in to the CIBERSORT website for details.