Deconvolute using EPIC and a custom signature matrix.

Deconvolute using EPIC and a custom signature matrix.

deconvolute_epic_custom(
  gene_expression_matrix,
  signature_matrix,
  signature_genes,
  genes_var = NULL,
  mrna_quantities = NULL,
  ...
)

Arguments

gene_expression_matrix

a m x n matrix with m genes and n samples

signature_matrix

a m x l matrix with m genes and l cell types. This matrix should contain the whole set of genes

signature_genes

a character vector of the gene names to use as signature needs to be smaller than the genes in the signature matrix

genes_var

(optional) a m x l matrix with m genes and l cell types, with the variability of each gene expression for each cell type. This will be used in the optimization

...

passed through to EPIC. A native argument takes precedence over an immunedeconv argument. See EPIC

mrna_cells

(optional) A named numeric vector with the amount of mRNA in arbitrary units for each of the reference cells and of the other uncharacterized cells.